refactor: split load-contract tests out of file_tests.rs (source-quality ceiling)
file_tests.rs reached 2518 lines against the 2500 ceiling after the
connectivity/determinism/error-reporting suites landed; the three newest
modules move to file_load_contract_tests.rs unchanged.
